% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction_protein.R
\name{batch_correction_protein}
\alias{batch_correction_protein}
\title{Calling different proteomics correction methods}
\usage{
batch_correction_protein(matrix, batch_map, batch, correction_method)
}
\arguments{
\item{correction_method}{}
}
\description{
Calling different proteomics correction methods
}
