% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{cell_type_enrichment}
\alias{cell_type_enrichment}
\title{Cell type enrichment of communities}
\usage{
cell_type_enrichment(multiassay, communities)
}
\arguments{
\item{communities}{}
}
\description{
Cell type enrichment of communities
}
