% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_TF.R
\name{circos_TF}
\alias{circos_TF}
\title{Creates a circos plot of identified TF and gene targets in modules}
\usage{
circos_TF(TF)
}
\arguments{
\item{TF}{output of \code{Transcription_Factor_enrichment()}}
}
\value{
circos plot
}
\description{
Creates a circos plot of identified TF and gene targets in modules
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
