% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_corrplot.R
\name{circular_corrplot}
\alias{circular_corrplot}
\title{Circular heatmap of module-trait correlations and module-cell type enrichments}
\usage{
circular_corrplot(moduleTraitCor, moduleColors, cell_type, names = NULL)
}
\arguments{
\item{data}{}
}
\description{
Circular heatmap of module-trait correlations and module-cell type enrichments
}
