% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_heatmap.R
\name{circular_heatmap}
\alias{circular_heatmap}
\title{Title}
\usage{
circular_heatmap(
  network_communities,
  cell_type_communities,
  data = "bundle_WGCNA_rna/rna-networkConstruction-auto_new.RData"
)
}
\arguments{
\item{data}{}
}
\description{
Title
}
