% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_DE.R, R/limma_DE.R
\name{clustering_DE_analysis}
\alias{clustering_DE_analysis}
\title{Performs differential expression between clusters}
\usage{
clustering_DE_analysis(
  normalized_data = multimodal$rna_processed,
  colData,
  dependent = "Cluster",
  levels = c("CS2", "CS1"),
  log2FoldChange = 0.5
)

clustering_DE_analysis(
  normalized_data = multimodal$rna_processed,
  colData,
  dependent = "Cluster",
  levels = c("CS2", "CS1"),
  log2FoldChange = 0.5
)
}
\arguments{
\item{log2FoldChange}{}
}
\description{
Performs differential expression between clusters

Title
}
