% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_correction.R
\name{combat_correction}
\alias{combat_correction}
\title{Proteomics normalisation batch correction using Combat}
\usage{
combat_correction(matrix, batch_map)
}
\arguments{
\item{batch_map}{}
}
\description{
Proteomics normalisation batch correction using Combat
}
