% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{communities_network}
\alias{communities_network}
\title{Detects communities in multi-omics network}
\usage{
communities_network(igraph, community_detection = "louvain")
}
\arguments{
\item{igraph}{Multi-omics network}

\item{community_detection}{Available methods include \code{edge_betweeness},
\code{leading_eigen},\code{walktrap},\code{fastgreedy},\code{louvain} all available from the
\code{igraph} package.}
}
\value{
List object including \code{community_object} and \code{communities}
}
\description{
Detects communities in multi-omics network
}
