% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_heatmap.R
\name{correlation_heatmap_clusters}
\alias{correlation_heatmap_clusters}
\title{Correlation heatmap between Factors and covariates of interest}
\usage{
correlation_heatmap_clusters(
  integrated_object,
  metadata,
  covariates = c("AD", "Braak", "amyloid", "pTau", "PHF1")
)
}
\arguments{
\item{integrated_object}{integrated object}

\item{metadata}{metadata}

\item{covariates}{covariates to be used for correlation}
}
\value{
Correlation heatmap
}
\description{
Correlation heatmap between Factors and covariates of interest
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{visualise_pathway}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
