% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rna.R
\name{counts_rna}
\alias{counts_rna}
\title{Adds custom processed dataset to multiassay object}
\usage{
counts_rna(multiassay, custom_processed_df)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{custom_processed_df}{custom processed data frame}
}
\value{
Adds custom processed dataset to Multiassay experiment object
generated by Omix
}
\description{
Adds custom processed dataset to multiassay object
}
\seealso{
Other Helper: 
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{format_res_deseq}()},
\code{\link{format_res_limma}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
