% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_multiomics.R, R/WIP_DE_multiomics.R
\name{DE_multiomics}
\alias{DE_multiomics}
\title{Performs DE on multiomics object from the processed slots}
\usage{
DE_multiomics(
  multiassay,
  dependent = "diagnosis",
  covariates = c("pmi", "msex", "age_death"),
  levels = c("Control", "Mid", "Late")
)

DE_multiomics(
  multiassay,
  dependent = "diagnosis",
  covariates = c("pmi", "msex", "age_death"),
  levels = c("Control", "Mid", "Late")
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{dependent}{Dependent variable for the differential expression analysis, usually the
disease group variable}

\item{covariates}{Technical and biological covariates. Set as NULL is data is denoised.}

\item{levels}{}

\item{log2FoldChange}{Log2FC absolute threshold. Default to 0.2}
}
\value{
List of differential expression results
}
\description{
Performs DE on multiomics object from the processed slots

Performs DE on multiomics object
}
