% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.communities_network}
\alias{.communities_network}
\title{Detects communities in multi-omics network}
\usage{
.communities_network(igraph, community_detection = "louvain")
}
\arguments{
\item{community_detection}{}
}
\description{
Detects communities in multi-omics network
}
