% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.community_graph}
\alias{.community_graph}
\title{Creates a subnetwork graph based on detected communities}
\usage{
.community_graph(igraph, community_object, community = 1)
}
\arguments{
\item{community}{}
}
\description{
Creates a subnetwork graph based on detected communities
}
