% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.dot_plot_enrichr_semantics}
\alias{.dot_plot_enrichr_semantics}
\title{Curates dotplot for \code{pathway_analysis_enrichr()} on a required semantics}
\usage{
.dot_plot_enrichr_semantics(dt, semantics = "microglial")
}
\arguments{
\item{dt}{functional enrichment results}

\item{semantics}{vector of biological terms}
}
\description{
Curates dotplot for \code{pathway_analysis_enrichr()} on a required semantics
}
