% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{.format_res_table_enrichr}
\alias{.format_res_table_enrichr}
\title{Format result table for \code{pathway_analysis_enrichr()}}
\usage{
.format_res_table_enrichr(res, min_overlap = 3)
}
\arguments{
\item{res}{Internal to \code{pathway_analysis_enrichr()}}

\item{min_overlap}{Default to 3}
}
\description{
Format result table for \code{pathway_analysis_enrichr()}
}
\keyword{internal}
