% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{.get_ID_names}
\alias{.get_ID_names}
\title{Convert nomenclature}
\usage{
.get_ID_names(
  multiassay,
  id,
  omic = "protein",
  from = "uniprot_id",
  to = "gene_name"
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{id}{Character vector to convert}

\item{omic}{Supported omic layers include \code{protein} or \code{rna}}

\item{from}{Possible nomenclatures include \code{uniprot_id} ,\code{ensembl_gene_id}, \code{gene_name}}

\item{to}{Possible nomenclatures include \code{uniprot_id} ,\code{ensembl_gene_id}, \code{gene_name}}
}
\value{
vector of new IDs
}
\description{
Convert nomenclature
}
