% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{.get_ID_type}
\alias{.get_ID_type}
\title{Get nomenclature type}
\usage{
.get_ID_type(character_vector)
}
\arguments{
\item{character_vector}{}
}
\value{
ID type between \code{entrez_gene_id},\code{gene_name},\code{ensembl_gene_id},
\code{uniprot_id}, \code{not_identified}
}
\description{
Get nomenclature type
}
