% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{.get_multimodal_object}
\alias{.get_multimodal_object}
\title{Extract the multimodal object}
\usage{
.get_multimodal_object(
  multiassay,
  slots = c("rna_processed", "protein_processed"),
  intersect_genes = FALSE,
  ID_type = "gene_name",
  ...
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{slots}{Default to processed slots}

\item{intersect_genes}{Logical whether to intersect common genes between
omic layers}

\item{ID_type}{Possible values are \code{gene_name}, \code{original},
Default to gene_name}
}
\value{
A list object containing the multimodal object and metadata
}
\description{
Extract the multimodal object
}
