% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_protein.R
\name{.impute_distribution}
\alias{.impute_distribution}
\title{Imputation of missing values based on distribution}
\usage{
.impute_distribution(df, width = 0.3, downshift = 1.8)
}
\arguments{
\item{downshift}{}
}
\description{
Imputation of missing values based on distribution
}
