% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.interactive_network}
\alias{.interactive_network}
\title{Get interactive network from igraph object}
\usage{
.interactive_network(igraph, communities, cluster)
}
\arguments{
\item{igraph}{}
}
\description{
Get interactive network from igraph object
}
