% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.multiomics_network}
\alias{.multiomics_network}
\title{Get multiomics networks from multiassay object}
\usage{
.multiomics_network(multiassay, list, correlation_threshold = 0.5)
}
\arguments{
\item{correlation_threshold}{}
}
\description{
Get multiomics networks from multiassay object
}
