% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.multiomics_network_cluster}
\alias{.multiomics_network_cluster}
\title{Creates a multi-omics network based on clustered subgroups following multi-omics clustering}
\usage{
.multiomics_network_cluster(
  multiassay,
  integration = "iCluster",
  cluster = 1,
  list,
  correlation_threshold = 0.5
)
}
\arguments{
\item{correlation_threshold}{}
}
\description{
Creates a multi-omics network based on clustered subgroups following multi-omics clustering
}
