% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{dotplot_enrichment_MOFA}
\alias{dotplot_enrichment_MOFA}
\title{Creates a functional enrichment dotplot}
\usage{
dotplot_enrichment_MOFA(
  enrichment.results,
  factor,
  alpha = 0.1,
  max.pathways = 25,
  text_size = 1,
  dot_size = 5
)
}
\arguments{
\item{dot_size}{}
}
\description{
Creates a functional enrichment dotplot
}
