% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{dotplot_enrichment_MOFA_semantics}
\alias{dotplot_enrichment_MOFA_semantics}
\title{Creates a functional enrichment dotplot based on a given semantics}
\usage{
dotplot_enrichment_MOFA_semantics(
  enrichment.results,
  semantics,
  factor,
  alpha = 0.05,
  max.pathways = 25,
  text_size = 1,
  dot_size = 5
)
}
\arguments{
\item{dot_size}{}
}
\description{
Creates a functional enrichment dotplot based on a given semantics
}
