% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment_custom}
\alias{enrichment_custom}
\title{Functional enrichment based on given reference annotations}
\usage{
enrichment_custom(genes, reference, genesets, adj = "fdr", verbose = FALSE)
}
\arguments{
\item{genes}{Gene list (vector)}

\item{reference}{Reference background (vector)}

\item{genesets}{List of genesets}

\item{adj}{Multiple testing adjustment method. Default to 'fdr'}

\item{verbose}{Default to FALSE}
}
\value{
Functional enrichment result
}
\description{
Functional enrichment based on given reference annotations
}
\seealso{
Other Enrichment analysis: 
\code{\link{GSEA}()},
\code{\link{cell_type_enrichment}()},
\code{\link{pathway_analysis_enrichr}()}
}
\concept{Enrichment analysis}
