% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{extract_loadings}
\alias{extract_loadings}
\title{Extract loadings}
\usage{
extract_loadings(
  integrated_object,
  metadata,
  component = 1,
  sense_check_variable = NULL,
  threshold = 0.2
)
}
\arguments{
\item{integrated_object}{integrated_object}

\item{metadata}{metadata}

\item{component}{component}

\item{sense_check_variable}{default to \code{NULL}. High weights should
coincide with stronger correlation if the sense_check_variable is an
important driver of variation in the designated factor. Will be used to
generate a plot if set to a covariate.}

\item{threshold}{Absolute threshold to filter weights}
}
\value{
List with positive and negative weights above and below the
specified threshold, distirbution plots of the weights, and dataframes
of the weights for each omic layer.
}
\description{
Extract loadings
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
