% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weights.R
\name{extract_weigths}
\alias{extract_weigths}
\title{Extract_weigths from MOFA model for a given factor of interest}
\usage{
extract_weigths(
  model,
  factor = 1,
  threshold = 0.3,
  variable,
  sense_check_variable
)
}
\arguments{
\item{variable}{}
}
\description{
Extract_weigths from MOFA model for a given factor of interest
}
