% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_protein.R
\name{filter_protein}
\alias{filter_protein}
\title{Data filtering function}
\usage{
filter_protein(matrix, groups, min_sample = 0.5, at_least_one = FALSE)
}
\arguments{
\item{matrix}{data frame containing LOG2 data}

\item{groups}{A character vector dictating the grouping}

\item{min_sample}{minimum \% of samples that should have non missing
protein value, or else the protein is excluded}

\item{at_least_one}{TRUE means to keep the row if min_count is met for at
least one condition, FALSE means min_count must be met across all
groups for retention}
}
\value{
Filtered protein matrix
}
\description{
Data filtering function
}
