% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{format_res_deseq}
\alias{format_res_deseq}
\title{Formatting differential expression results from DESEQ2}
\usage{
format_res_deseq(
  dt,
  log2FoldChange = 0,
  padj = 0.05,
  n_label = 10,
  gene_id_conversion = NULL,
  filter_protein_coding = TRUE
)
}
\arguments{
\item{filter_protein_coding}{}
}
\description{
Formatting differential expression results from DESEQ2
}
