% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{format_res_limma}
\alias{format_res_limma}
\title{Title}
\usage{
format_res_limma(
  dt,
  gene_id_conversion = NULL,
  log2FoldChange = 0,
  n_label = 10,
  padj = 0.05,
  ylim = c(0, 10)
)
}
\arguments{
\item{ylim}{}
}
\description{
Title
}
