% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mae.R
\name{generate_mae}
\alias{generate_mae}
\title{Generate a MultiAssayExperiment object from single-omics data}
\usage{
generate_mae(
  rawdata_rna,
  rawdata_protein,
  map_rna,
  map_protein,
  metadata_rna,
  metadata_protein,
  sample_metadata,
  map_by_column,
  rna_id_type,
  protein_id_type,
  rna_qc_data = TRUE,
  rna_qc_data_matrix
)
}
\arguments{
\item{rawdata_rna}{A data-frame containing raw RNA count where rows and
columns correspond to genes and samples respectively}

\item{rawdata_protein}{A data-frame containing raw protein abundance where
rows and columns correspond to genes and samples respectively}

\item{map_rna}{A data-frame of two columns named \code{primary} and \code{colname}
where primary should contain unique sample name with a link to sample
metadata and colname is the column names of the \code{rawdata_rna} data-frame}

\item{map_protein}{A data-frame of two columns named \code{primary} and \code{colname}
where primary should contain unique sample name with a link to sample
metadata and colname is the column names of the \code{rawdata_protein} data-frame}

\item{metadata_rna}{A data-frame containing \code{rna} assay specific metadata
where rownames are same as the colnames of \code{rawdata_rna} data.frame}

\item{metadata_protein}{A data-frame containing \code{protein} assay specific
metadata where rownames are same as the colnames of \code{rawdata_protein}
data.frame}

\item{sample_metadata}{A data-frame containing sample level metadata for
both omics assays}

\item{map_by_column}{The common column name to link \code{metadata_rna}
and \code{metadata_protein} to the \code{sample_metadata}}

\item{rna_id_type}{The gene ID type used for the rownames of \code{rawdata_rna}.
Possible values are \code{ensembl_gene_id}, \code{gene_name}.}

\item{protein_id_type}{The protein ID type used for the rownames of
\code{rawdata_protein}. Possible values are \code{uniprot_id}, \code{gene_name}.}

\item{rna_qc_data}{Logical whether to add rna QC data.}

\item{rna_qc_data_matrix}{A data.frame containing RNA qc level data where
rownames are same as the colnames of \code{rawdata_rna} data.frame.}
}
\value{
a MultiAssayExperiment object
}
\description{
Generate a MultiAssayExperiment object from single-omics data matrix
and metadata. Currently supports transcriptomics and proteomics data only.
}
