% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multiassay.R
\name{generate_multiassay}
\alias{generate_multiassay}
\title{Generate a MultiAssayExperiment object from multiple single-omics data}
\usage{
generate_multiassay(
  rawdata_rna,
  rawdata_protein,
  individual_to_sample = FALSE,
  map_rna,
  map_protein,
  metadata_rna,
  metadata_protein,
  individual_metadata,
  map_by_column,
  rna_qc_data = FALSE,
  rna_qc_data_matrix,
  organism = "human"
)
}
\arguments{
\item{rawdata_rna}{A data-frame containing raw RNA count where rows and
columns correspond to genes and samples respectively}

\item{rawdata_protein}{A data-frame containing raw protein abundance where
rows and columns correspond to genes and samples respectively}

\item{individual_to_sample}{Logical whether individual ID and sample names in
raw data matrices are the same. If they are different, \code{mpa_rna} and
\code{map_protein}
dataframes should be provided. Default to \code{FALSE}.}

\item{map_rna}{A data-frame of two columns named \code{primary} and \code{colname}
where primary should contain unique ID name with a link to individual
metadata and colname is the column names of the \code{rawdata_rna} data-frame}

\item{map_protein}{A data-frame of two columns named \code{primary} and \code{colname}
where primary should contain unique ID name with a link to individual
metadata and colname is the column names of the \code{rawdata_protein} data-frame}

\item{metadata_rna}{A data-frame containing \code{rna} assay specific metadata
where rownames are same as the colnames of \code{rawdata_rna} data.frame}

\item{metadata_protein}{A data-frame containing \code{protein} assay specific
metadata where rownames are same as the colnames of \code{rawdata_protein}
data.frame}

\item{individual_metadata}{A data-frame containing sample level metadata}

\item{map_by_column}{The common column name to link \code{metadata_rna}
and \code{metadata_protein} to the \code{individual_metadata}}

\item{rna_qc_data}{Logical whether to add rna QC data.}

\item{rna_qc_data_matrix}{A data.frame containing RNA qc level data where
rownames are same as the colnames of \code{rawdata_rna} data.frame.}

\item{organism}{The organism type for transcripts and proteins.
Possible values are \code{human} and \code{mouse}. Default to \code{human}.}
}
\value{
a MultiAssayExperiment object
}
\description{
Generate a MultiAssayExperiment object from single-omics data matrix
and metadata. Currently supports transcriptomics and proteomics data only.
}
\seealso{
Other Pre-processing: 
\code{\link{batch_correction_protein}()},
\code{\link{combat_correction}()},
\code{\link{filter_protein}()},
\code{\link{process_protein}()},
\code{\link{process_rna}()}
}
\concept{Pre-processing}
