% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_background}
\alias{get_background}
\title{Get background genes}
\usage{
get_background(multiassay, of = "full")
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{of}{Possible values are \code{full}, \code{protein},\code{rna}, respectively for
the background of the union between transcriptomics and proteomics, or of one
of the omic layers
default to \code{full}.}
}
\value{
Background genes character vector
}
\description{
Get background genes
}
\seealso{
Other Helper: 
\code{\link{ask_chatgpt}()},
\code{\link{counts_rna}()},
\code{\link{format_res_deseq}()},
\code{\link{format_res_limma}()},
\code{\link{get_metadata}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
