% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(multiassay, omic)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{omic}{Supported omic layers include \code{protein} or \code{rna}}
}
\value{
colData
}
\description{
Get metadata
}
\seealso{
Other Helper: 
\code{\link{ask_chatgpt}()},
\code{\link{counts_rna}()},
\code{\link{format_res_deseq}()},
\code{\link{format_res_limma}()},
\code{\link{get_background}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
