% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_module_annotation}
\alias{get_module_annotation}
\title{Get module functional annotation}
\usage{
get_module_annotation(gene_list, moduleColors)
}
\arguments{
\item{gene_list}{Character vector of genes belonging to a module}

\item{moduleColors}{from WGCNA}
}
\value{
Functional annotation of modules
}
\description{
Get module functional annotation
}
