% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_clustering_heatmap.R
\name{getMoHeatmap}
\alias{getMoHeatmap}
\alias{multiomics_clustering_heatmap}
\title{Get multi-omics comprehensive heatmap}
\usage{
multiomics_clustering_heatmap(
  data = NULL,
  is.binary = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  row.title = c("Data1", "Data2", "Data3", "Data4", "Data5", "Data6"),
  legend.name = c("Data1", "Data2", "Data3", "Data4", "Data5", "Data6"),
  clust.res = NULL,
  clust.dend = NULL,
  show.col.dend = TRUE,
  show.colnames = FALSE,
  show.row.dend = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  show.rownames = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  clust.dist.row = c("pearson", "pearson", "pearson", "pearson", "pearson", "pearson"),
  clust.method.row = c("ward.D", "ward.D", "ward.D", "ward.D", "ward.D", "ward.D"),
  clust.col = c("#2EC4B6", "#E71D36", "#FF9F1C", "#BDD5EA", "#FFA5AB", "#011627",
    "#023E8A", "#9D4EDD"),
  color = rep(list(c("#00FF00", "#000000", "#FF0000")), length(data)),
  annCol = NULL,
  annColors = NULL,
  annRow = NULL,
  width = 6,
  height = 4,
  fig.path = getwd(),
  fig.name = "moheatmap"
)
}
\arguments{
\item{data}{A list of data frame or matrix storing multiple omics data with rows for features and columns for samples.}

\item{is.binary}{A logicial vector to indicate if the subdata is binary matrix of 0 and 1 such as mutation.}

\item{row.title}{A string vector to assign titles for each subdata.}

\item{legend.name}{A string vector to assign legend title for each subdata.}

\item{clust.res}{A clust.res object returned by \code{getMOIC()} with one specified algorithm or \verb{get\\\%algorithm_name\\\%} or \code{getConsensusMOIC()} with a list of multiple algorithms.}

\item{clust.dend}{A dendrogram object returned returned by \code{getMOIC()} with one specified algorithm or \verb{get\\\%algorithm_name\\\%} or \code{getConsensusMOIC()} with a list of multiple algorithms.}

\item{show.col.dend}{A logical vector to indicate if showing the dendrogram for column at the top of heatmap.}

\item{show.colnames}{A logical vector to indicate if showing the names for column at the bottom of heatmap.}

\item{show.row.dend}{A logical vector to indicate if showing the dendrogram for row of each subdata.}

\item{show.rownames}{A logical vector to indicate if showing the names for row of each subdata.}

\item{clust.dist.row}{A string vector to assign distance method for clustering each subdata at feature dimension.}

\item{clust.method.row}{A string vector to assign clustering method for clustering each subdata at feature dimension.}

\item{clust.col}{A string vector storing colors for annotating each subtype at the top of heatmap.}

\item{color}{A list of string vectors storing colors for each subheatmap of subdata.}

\item{annCol}{A data.frame storing annotation information for samples with exact the same sample order with data parameter.}

\item{annColors}{A list of string vectors for colors matched with annCol.}

\item{annRow}{A list of string vectors to indicate which features belong to which subdata should be annotated specifically in subheatmap.}

\item{width}{An integer value to indicate the width for each subheatmap with unit of cm.}

\item{height}{An integer value to indicate the height for each subheatmap with unit of cm.}

\item{fig.path}{A string value to indicate the output path for storing the comprehensive heatmap.}

\item{fig.name}{A string value to indicate the name of the comprehensive heatmap.}
}
\value{
A pdf of multi-omics comprehensive heatmap
}
\description{
This function vertically concatenates multiple heatmap derived from each omics data. \code{getMoHeatmap} supports customized column annotation and is able to mark the selected features if indicated.
}
\references{
Gu Z, Eils R, Schlesner M (2016). Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
\seealso{
Other Plotting: 
\code{\link{circular_heatmap}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
