% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_DIABLO}
\alias{integrate_with_DIABLO}
\title{Vertical integration with DIABLO including model tuning}
\usage{
integrate_with_DIABLO(
  multimodal_omics,
  Y,
  ncomp,
  design,
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10)),
  list.keepX
)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{get_multimodal_object}}

\item{Y}{Outcome variable}

\item{ncomp}{Number of components}

\item{design}{Available options are cor, full. Possible designs for
integration are \code{cor}(strength of correlation between PC1 of omics layers)
or \code{full} (The strength of all relationships between dataframes is
maximised (= 1) – a “fully connected” design)}

\item{range}{a named list specifying the number of rna and protein.}

\item{list.keepX}{list to keep rows}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with DIABLO including model tuning
}
\seealso{
Other Multi-omic integration: 
\code{\link{integrate_with_MBPLS}()},
\code{\link{integrate_with_MEIFESTO}()},
\code{\link{integrate_with_MOFA}()},
\code{\link{integrate_with_sMBPLS}()},
\code{\link{vertical_integration}()}
}
\concept{Multi-omic integration}
