% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_iCluster}
\alias{integrate_with_iCluster}
\title{Vertical integration with iCluster}
\usage{
integrate_with_iCluster(multimodal_omics, try.N.clust = 2:4)
}
\arguments{
\item{try.N.clust}{}
}
\description{
Vertical integration with iCluster
}
