% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_MBPLS}
\alias{integrate_with_MBPLS}
\title{Vertical integration with MBPLS}
\usage{
integrate_with_MBPLS(
  multimodal_omics,
  Y,
  design = c("cor", "full", "avg"),
  ncomp
)
}
\arguments{
\item{ncomp}{}
}
\description{
Vertical integration with MBPLS
}
