% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_MEIFESTO}
\alias{integrate_with_MEIFESTO}
\title{Vertical integration with MEIFESTO}
\usage{
integrate_with_MEIFESTO(
  multimodal_omics,
  num_factors = 5,
  scale_views = T,
  metadata,
  time = "pseudotime"
)
}
\arguments{
\item{time}{}
}
\description{
Vertical integration with MEIFESTO
}
