% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_MOFA}
\alias{integrate_with_MOFA}
\title{Vertical integration with MOFA}
\usage{
integrate_with_MOFA(multimodal_omics, num_factors = 5, scale_views = T)
}
\arguments{
\item{scale_views}{}
}
\description{
Vertical integration with MOFA
}
