% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_sMBPLS}
\alias{integrate_with_sMBPLS}
\title{Vertical integration with SMBPLS}
\usage{
integrate_with_sMBPLS(
  multimodal_omics,
  Y,
  design = c("cor", "full", "avg"),
  ncomp,
  list.keepX = list(mRNA = c(50), proteins = c(50))
)
}
\arguments{
\item{list.keepX}{}
}
\description{
Vertical integration with SMBPLS
}
