% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_enrichment_multiGSEA.R
\name{integrated_enrichment_mutliGSEA}
\alias{integrated_enrichment_mutliGSEA}
\title{Integrated pathway enrichment with multiGSEA}
\usage{
integrated_enrichment_mutliGSEA(
  integration = "pvalue",
  DE_comparison_dataframe = NULL,
  weights = NULL,
  databases = "reactome",
  additional_database_gmt = FALSE
)
}
\arguments{
\item{integration}{Options are pvalue padj weights}

\item{DE_comparison_dataframe}{if integration == pvalue or padj, provide the differential expression comparison dataframe generated with \code{single_omic_comparisons()}}

\item{weights}{if integration == weights, provide weights dataframe generated with \code{extract_weights()}}

\item{databases}{Options are: all kegg panther pathbank pharmgkb reactome smpdb wikipathways or a combination thereof. Use NULL for GMT file only}

\item{additional_database_gmt}{provide GMT file path}
}
\value{
List of results
}
\description{
Integrated pathway enrichment with multiGSEA
}
\seealso{
Other Enrichment analysis: 
\code{\link{GSEA}()},
\code{\link{cell_type_enrichment}()},
\code{\link{enrichment_custom}()},
\code{\link{pathway_analysis_enrichr}()}
}
\concept{Enrichment analysis}
