% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_clustering.R
\name{integrative_results_clustering}
\alias{integrative_results_clustering}
\title{Downstream analysis for integrated multi-omics clustering}
\usage{
integrative_results_clustering(
  multiassay,
  integration = "iCluster",
  enrichment_method,
  annotations_continuous,
  annotations_cat,
  res.path,
  correlation_threshold = 0.5,
  geneset = NULL
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{integration}{Possible multiomics clustering include \code{iCluster}}

\item{enrichment_method}{Available methods include \code{enrichGO} and \code{enrichr}. Default to \code{enrichr}.}

\item{annotations_continuous}{continuous covariates}

\item{annotations_cat}{categorical covariates}

\item{correlation_threshold}{Absolute correlation threshold to draw edges in the network}

\item{geneset}{Default to NULL}
}
\value{
List object of integrated results
}
\description{
Downstream analysis for integrated multi-omics clustering
}
