% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_unsupervised.R
\name{integrative_results_unsupervised}
\alias{integrative_results_unsupervised}
\title{Downstream analysis for integrated multi-omics data}
\usage{
integrative_results_unsupervised(
  multiassay,
  integration = "MOFA",
  dependent = "diagnosis",
  correlation_threshold = 0.5,
  disease_id = "MONDO_0004975",
  covariates = "diagnosis"
)
}
\arguments{
\item{covariates}{}
}
\description{
Downstream analysis for integrated multi-omics data
}
