% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{interactive_network}
\alias{interactive_network}
\title{Get interactive network from igraph object}
\usage{
interactive_network(igraph, communities = TRUE, cluster)
}
\arguments{
\item{igraph}{Multi-omics network}

\item{communities}{Logical as whether to show communities on graph}

\item{cluster}{If commmunities = TRUE, provide a community object generated
by  \code{communities_network}function}
}
\value{
plots \code{visNetwork} graph
}
\description{
Get interactive network from igraph object
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{visualise_pathway}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
