% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_unsupervised.R
\name{MOFA_get_relevant_factor}
\alias{MOFA_get_relevant_factor}
\title{Title}
\usage{
MOFA_get_relevant_factor(MOFAobject, covariate)
}
\arguments{
\item{covariate}{}
}
\description{
Title
}
