% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_omic_report.R
\name{multi_omic_report}
\alias{multi_omic_report}
\title{Generate a report for single omic analysis}
\usage{
multi_omic_report(
  multiassay,
  report_folder_path = getwd(),
  report_file = "multi_omic_report_Omix",
  integration_model = "MOFA",
  database = "Reactome_2016",
  cluster = "C2-C1"
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{report_folder_path}{report_folder_path folder path to save the report.}

\item{report_file}{filename for report (without an extension).}

\item{integration_model}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{database}{Enrichment database \code{GO_Molecular_Function_2021},\code{GO_Cellular_Component_2021},
\code{GO_Biological_Process_2021}, \code{Reactome_2016} , \code{KEGG_2021_Human} , \code{MSigDB_Hallmark_2020}}

\item{cluster}{In the case of multi-omics clustering, choose two clusters of interest for differential analysis}
}
\value{
Multi-omic analysis report html
}
\description{
Generate a report for single omic analysis
}
\seealso{
Other Report: 
\code{\link{multi_omics_qc}()},
\code{\link{qc_report}()},
\code{\link{single_omic_report}()}
}
\concept{Report}
