% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_omics_qc.R
\name{multi_omics_qc}
\alias{multi_omics_qc}
\title{Generate multi-omics QC results}
\usage{
multi_omics_qc(multiassay)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}
}
\value{
parameters$processing_outputs$multiomics slot
}
\description{
Generate multi-omics QC results
}
\seealso{
Other Report: 
\code{\link{multi_omic_report}()},
\code{\link{pathway_report}()},
\code{\link{proteomics_report}()},
\code{\link{qc_report}()},
\code{\link{single_omic_report}()}
}
\concept{Report}
