% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_heatmap.R
\name{multiomics_heatmap}
\alias{multiomics_heatmap}
\title{Multi-omics heatmap}
\usage{
multiomics_heatmap(
  multimodal_omics,
  metadata,
  integration_model = "iCluster",
  covariates = c("PHF1", "diagnosis"),
  omic = c("rna", "protein"),
  weights,
  de_features,
  direction = c("both", "positive", "negative"),
  cluster = NULL,
  num_features = NULL
)
}
\arguments{
\item{multimodal_omics}{multimodal_omics list}

\item{metadata}{metdata dataframe}

\item{integration_model}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{covariates}{clinical covariates}

\item{omic}{\code{rna},\code{protein} or \code{both}}

\item{weights}{weights from integrative results}

\item{de_features}{If integration model is clustering, use differentially expressed features}

\item{direction}{\code{positive} or \code{negative}}

\item{cluster}{default to NULL}

\item{num_features}{default to NULL will plot all the selected features. Otherwise numerical variable to define the maximum number of features}
}
\value{
Heatmap
}
\description{
Multi-omics heatmap
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
