% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_modules.R
\name{multiomics_modules}
\alias{multiomics_modules}
\title{Multi-omics modules downstream analysis including eigenvalue and correlation to clinical traits}
\usage{
multiomics_modules(
  multiassay,
  metadata,
  covariates = c("pseudotime", "PHF1", "amyloid", "pTau"),
  communities
)
}
\arguments{
\item{communities}{}
}
\description{
Multi-omics modules downstream analysis including eigenvalue and correlation to clinical traits
}
