% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{multiomics_network_cluster}
\alias{multiomics_network_cluster}
\title{Creates a multi-omics network based on clustered subgroups following
multi-omics clustering}
\usage{
multiomics_network_cluster(
  multiassay,
  integration = "iCluster",
  cluster = 1,
  list,
  correlation_threshold = 0.5
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration}{multi-omics clustering integration method.
Default \code{iCluster}}

\item{cluster}{Cluster of interest}

\item{list}{Named list of processed omic slots and features of interest}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}
}
\value{
\code{Igraph} object
}
\description{
Creates a multi-omics network based on clustered subgroups following
multi-omics clustering
}
