% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{multiomics_network_matrix}
\alias{multiomics_network_matrix}
\title{Get multiomics networks from multimodal object}
\usage{
multiomics_network_matrix(multimodal_object, list, correlation_threshold = 0.5)
}
\arguments{
\item{multimodal_object}{multimodal_object generated by \code{get_multimodal_object}}

\item{list}{Named list of processed omic slots and features of interest}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}
}
\value{
List containing \code{graph} and \code{matrix} slots
}
\description{
Get multiomics networks from multimodal object
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{circos_TF}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
