% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{pathway_analysis_enrichr}
\alias{pathway_analysis_enrichr}
\title{Functional enrichment analysis using enrichR}
\usage{
pathway_analysis_enrichr(
  interest_gene = NULL,
  project_name = NULL,
  enrichment_database = c("GO_Molecular_Function_2021", "GO_Cellular_Component_2021",
    "GO_Biological_Process_2021", "WikiPathways_2021_Human", "Reactome_2016",
    "KEGG_2021_Human", "MSigDB_Hallmark_2020", "BioCarta_2016"),
  is_output = FALSE,
  output_dir = ".",
  min_overlap = 3,
  plot_n = 20
)
}
\arguments{
\item{enrichment_database}{Name of the database for enrichment. User can
specify one or more database names from \code{\link[enrichR:listEnrichrDbs]{enrichR::listEnrichrDbs()}}.}

\item{is_output}{If TRUE a folder will be created and results of enrichment
analysis will be saved otherwise a R list will be returned. Default FALSE}

\item{output_dir}{Path for the output directory. Default is current dir.}

\item{interest_genes}{vector containing genes of interest
Column names should be gene, logFC, pval and padj respectively.}
}
\value{
enrichment_result a list of data.frames containing enrichment output
and a list of plots of top 10 significant genesets.
}
\description{
Performs impacted pathway analysis with a list of genes.
}
\seealso{
Other Enrichment analysis: 
\code{\link{GSEA}()},
\code{\link{cell_type_enrichment}()},
\code{\link{enrichment_custom}()}
}
\concept{Enrichment analysis}
