% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_report.R
\name{pathway_report}
\alias{pathway_report}
\title{Generate a report for single omic analysis}
\usage{
pathway_report(
  pathways,
  report_folder_path = getwd(),
  report_file = "single_omic_report_Omix",
  database = "Reactome_2016",
  num_path = 20
)
}
\arguments{
\item{pathways}{list of pathways}

\item{report_folder_path}{report_folder_path folder path to save the report.}

\item{report_file}{filename for report (without an extension).}

\item{database}{Enrichment database \code{GO_Molecular_Function_2021},\code{GO_Cellular_Component_2021},
\code{GO_Biological_Process_2021}, \code{Reactome_2016} , \code{KEGG_2021_Human} , \code{MSigDB_Hallmark_2020}}
}
\value{
Single omic analyses report html
}
\description{
Generate a report for single omic analysis
}
\seealso{
Other Report: 
\code{\link{multi_omic_report}()},
\code{\link{multi_omics_qc}()},
\code{\link{proteomics_report}()},
\code{\link{qc_report}()},
\code{\link{single_omic_report}()}
}
\concept{Report}
