% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{plot_communities}
\alias{plot_communities}
\title{Plots multi-ommics network communities}
\usage{
plot_communities(igraph, community_object)
}
\arguments{
\item{igraph}{Multi-omics network}

\item{community_object}{community_object}
}
\value{
multi-omics network plot
}
\description{
Plots multi-ommics network communities
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
