% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{plot_components}
\alias{plot_components}
\title{Plot components}
\usage{
plot_components(
  integrated_object,
  metadata,
  dependent,
  layer = c("mRNA", "proteins")
)
}
\arguments{
\item{integrated_object}{integrated object}

\item{metadata}{metadata}

\item{dependent}{dependent variable}

\item{layer}{"mRNA", "proteins"}
}
\value{
Plot
}
\description{
Plot components
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{visualise_pathway}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
