% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{plot_geneset_clusters}
\alias{plot_geneset_clusters}
\title{Plots FGSEA results}
\usage{
plot_geneset_clusters(gs_results, myGO, min.sz = 4, main = "GSEA clusters")
}
\arguments{
\item{main}{}
}
\description{
Plots FGSEA results
}
