% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot loadings}
\usage{
plot_loadings(
  loadings = loadings$weights_df,
  layer = "rna",
  sign = NULL,
  abs = TRUE,
  nfeatures = 30
)
}
\arguments{
\item{loadings}{loadings}

\item{layer}{\code{rna} or \code{proteins}}

\item{sign}{\code{positive} or \code{negative}}

\item{abs}{TRUE}

\item{nfeatures}{number of features to plot}
}
\value{
Plot of top feature loadings
}
\description{
Plot loadings
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{visualise_pathway}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
