% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{plot_OpenTarget}
\alias{plot_OpenTarget}
\title{Interface to OpenTargets database}
\usage{
plot_OpenTarget(disease_id = "MONDO_0004975", size = 1000, genes)
}
\arguments{
\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease) Different disease ontologies can be found on
https://www.ebi.ac.uk/ols/ontologies}

\item{size}{Default to 3000}

\item{genes}{input gene vector}
}
\value{
ggplot showing association scores from OpenTargets
}
\description{
Interface to OpenTargets database
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
