% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_protein.R
\name{process_protein}
\alias{process_protein}
\title{Process bulk proteomics data}
\usage{
process_protein(
  multiassay,
  filter = TRUE,
  min_sample = 0.5,
  dependent = "diagnosis",
  levels = c("Control", "AD"),
  imputation = "minimum_value",
  remove_feature_outliers = TRUE,
  batch_correction = TRUE,
  batch = "batch",
  correction_method = c("Combat", "Median_centering"),
  remove_sample_outliers = TRUE,
  denoise = FALSE,
  covariates = c("sex", "age")
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{filter}{Logical whether to perform protein filtering}

\item{min_sample}{Percentage of samples that must non missing protein
value, or the protein is filtered out
protein value, or else the protein is excluded}

\item{dependent}{Dependent variable for groups.
If no specific dependent variable exist, set parameter as NULL}

\item{levels}{Character vector with reference group as first element.
Set parameter as NULL if dependent is NULL.}

\item{imputation}{Possible values are \code{distribution}, \code{minimum_value},\code{zero}
default to \code{minimum_value}.}

\item{remove_feature_outliers}{Logical whether to remove feature outliers}

\item{batch_correction}{Logical whether to perform  technical batch
correction}

\item{batch}{Technical batch}

\item{correction_method}{Correction methods available to counteract batch
effects. Possible values are \code{Combat} or \code{median_centering}}

\item{remove_sample_outliers}{Logical whether to remove sample outliers}

\item{denoise}{Logical whether to denoise biological covariates}

\item{covariates}{Covariates used in linear mixed model for denoising}
}
\value{
a MultiAssayExperiment object with \code{protein_processed} slot
}
\description{
Process bulk proteomics data
}
\seealso{
Other Pre-processing: 
\code{\link{batch_correction_protein}()},
\code{\link{combat_correction}()},
\code{\link{filter_protein}()},
\code{\link{generate_multiassay}()},
\code{\link{median_centering_correction}()},
\code{\link{process_rna}()}
}
\concept{Pre-processing}
