% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rna.R
\name{process_rna}
\alias{process_rna}
\title{Process batch transcriptomics data}
\usage{
process_rna(
  multiassay,
  transformation = c("vst", "rlog", "log2"),
  protein_coding,
  filter,
  min_count = 10,
  min_sample = 1,
  dependent,
  levels = NULL,
  covariates,
  batch_correction,
  batch,
  remove_sample_outliers
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{protein_coding}{Logical whether to filter out non protein coding genes}

\item{filter}{Logical whether to perform gene filtering}

\item{min_count}{Minimum count}

\item{min_sample}{Percentage of samples that have at least the \code{min_count},
or else the gene is filtered out}

\item{dependent}{Dependent variable for the DeSEQ2 analysis, usually the
disease group variable}

\item{levels}{Character vector with reference group as first element.
Set parameter as NULL if dependent is NULL.}

\item{covariates}{Technical and biological covariates to use for
DESEQ2 analysis and denoising}

\item{batch_correction}{Logical whether to perform batch correction. The
processed matrix returned will be denoised for designated covariates and
technical batch.}

\item{batch}{Technical batch, use batch2 for another technical batch}

\item{remove_sample_outliers}{Logical whether to remove sample outliers}
}
\value{
a MultiAssayExperiment object with processed slots
}
\description{
Process batch transcriptomics data
}
