% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_protein.R
\name{processed_proteomics}
\alias{processed_proteomics}
\title{Adds a custom processed proteomics dataset to multiassay object}
\usage{
processed_proteomics(multiassay, custom_processed_df)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{custom_processed_df}{processed dataframe}
}
\value{
multiassay object with custom slot
}
\description{
Adds a custom processed proteomics dataset to multiassay object
}
\seealso{
Other Helper: 
\code{\link{ask_chatgpt}()},
\code{\link{counts_rna}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{format_res_deseq}()},
\code{\link{format_res_limma}()},
\code{\link{get_background}()},
\code{\link{get_metadata}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
