% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rna.R
\name{processed_rna}
\alias{processed_rna}
\title{Adds a custom processed transcriptomics slot to multiassay object}
\usage{
processed_rna(multiassay, custom_processed_df)
}
\arguments{
\item{custom_processed_df}{}
}
\description{
Adds a custom processed transcriptomics slot to multiassay object
}
