% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_DE_analysis.R
\name{protein_DE_analysis}
\alias{protein_DE_analysis}
\title{Performs single omic analysis of proteomic layer.}
\usage{
protein_DE_analysis(
  multiassay,
  slot = "protein_processed",
  dependent = "diagnosis",
  covariates = NULL,
  levels = c("Control", "AD"),
  log2FoldChange = 0.5,
  padj = 0.5
)
}
\arguments{
\item{reference}{}
}
\description{
Performs single omic analysis of proteomic layer.
}
