% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_single_analysis.R
\name{protein_single_analysis}
\alias{protein_single_analysis}
\title{Title}
\usage{
protein_single_analysis(
  multiassay,
  slot = "protein_processed",
  dependent = "diagnosis",
  covariates = c("pmi", "msex", "age_death"),
  levels = c("Control", "Mid", "Late")
)
}
\arguments{
\item{reference}{}
}
\description{
Title
}
