% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_inference.R
\name{pseudotime_inference}
\alias{pseudotime_inference}
\title{Pseudotime_inference based on integrated embeddings and clusters}
\usage{
pseudotime_inference(
  model,
  clusters,
  time_factor = 8,
  second_factor = 1,
  pseudotime_var = "pseudotime",
  ...
)
}
\arguments{
\item{second_factor}{}
}
\description{
Pseudotime_inference based on integrated embeddings and clusters
}
