% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_ordering_modules.R
\name{pseudotime_ordering_modules}
\alias{pseudotime_ordering_modules}
\title{Generates a plot relating modules eigenvalue along pseudotime}
\usage{
pseudotime_ordering_modules(
  metadata,
  modules,
  variable,
  pseudotime_var = "pseudotime"
)
}
\arguments{
\item{pseudotime_var}{}
}
\description{
Generates a plot relating modules eigenvalue along pseudotime
}
