% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{qc_report}
\alias{qc_report}
\title{Generate a report for single omic analysis}
\usage{
qc_report(
  multiassay,
  report_folder_path = getwd(),
  report_file = "QC_report_Omix",
  clinical_covariates = c("age", "sex", "diagnosis", "Braak", "amyloid"),
  dependent = "diagnosis",
  map_by_column = "sample_id"
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{report_folder_path}{report_folder_path folder path to save the report.}

\item{report_file}{filename for report (without an extension).}

\item{clinical_covariates}{clinical covariates}

\item{dependent}{Dependent variable, usually the
disease group variable}

\item{map_by_column}{The common column name to link \code{metadata_rna}
and \code{metadata_protein} to the \code{individual_metadata}}
}
\value{
QC report html
}
\description{
Generate a report for single omic analysis
}
\seealso{
Other Report: 
\code{\link{multi_omic_report}()},
\code{\link{multi_omics_qc}()},
\code{\link{pathway_report}()},
\code{\link{proteomics_report}()},
\code{\link{single_omic_report}()}
}
\concept{Report}
