% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_DE_analysis.R
\name{rna_DE_analysis}
\alias{rna_DE_analysis}
\title{Performs single omic analysis of transcriptomics layer.}
\usage{
rna_DE_analysis(
  multiassay,
  dependent = "diagnosis",
  levels = NULL,
  filter_protein_coding = TRUE,
  log2FoldChange = 0.5
)
}
\arguments{
\item{multiassay}{Multiassay object generated by Omix}

\item{dependent}{Dependent variable in the differential expression analysis}

\item{levels}{Levels for dependent variable. Control group should be put
first.}

\item{filter_protein_coding}{Logical whether to filter out non protein coding
genes}

\item{log2FoldChange}{log2 fold-change cutoff.}
}
\value{
Adds differential expression results to Multiassay object generated
by Omix
}
\description{
Performs single omic analysis of transcriptomics layer.
}
\seealso{
Other Single-omic: 
\code{\link{DE_multiomics}()},
\code{\link{protein_DE_analysis}()},
\code{\link{single_omic_comparisons}()}
}
\concept{Single-omic}
