% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_single_analysis.R
\name{rna_single_analysis}
\alias{rna_single_analysis}
\title{Performs single omic analysis of transcriptomics layer.}
\usage{
rna_single_analysis(
  multiassay,
  dependent = "diagnosis",
  levels = NULL,
  filter_protein_coding = TRUE
)
}
\arguments{
\item{multiassay}{Multiassay object genereted by Omix}

\item{dependent}{Dependent variable in the differential expression analysis}

\item{levels}{Levels for dependent variable. Control group should be put
first.}

\item{fiter_protein_coding}{Logical whether to filter out non protein coding
genes}
}
\description{
Performs single omic analysis of transcriptomics layer.
}
