% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{sign_signature}
\alias{sign_signature}
\title{Extract sign}
\usage{
sign_signature(
  integrated_object,
  metadata,
  covariate = "pTau",
  relevant_component = "comp1"
)
}
\arguments{
\item{integrated_object}{integrated object}

\item{metadata}{metadata}

\item{covariate}{covariates}

\item{relevant_component}{component}
}
\value{
Sign
}
\description{
Extract sign
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{circos_TF}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()}
}
\concept{Multi-omic integration downstream analysis}
