% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_omic_comparison.R
\name{single_omic_comparison}
\alias{single_omic_comparison}
\title{Compare differential expressed genes and proteins}
\usage{
single_omic_comparison(multiassay, slot = "ADvsControl")
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{slot}{If more than two comparison groups designates the slot name of
interest in \code{names(multiassay@metadata$DEG)} and
\code{names(multiassay@metadata$DEP)}}
}
\description{
Compare differential expressed genes and proteins
}
