% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_omic_comparison.R
\name{single_omic_comparisons}
\alias{single_omic_comparisons}
\title{Compare differential expressed genes and proteins}
\usage{
single_omic_comparisons(
  multiassay,
  threshold = 0.05,
  pvalue = c("padj", "pvalue", "all"),
  filtering_options = "both_significant",
  additional_database_gmt = NULL
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{threshold}{is the pvalue threshold. Default to 0.05}

\item{pvalue}{Comparison reference.can be set to \code{pvalue} or \code{padj}, to consider
the normal p values or the ones after multiple testing correction,
respectively. With pvalue == \code{all}, no p value threshold is applied and all genes will be
compared. all genes are kept and only the direction is compared.}

\item{filtering_options}{default to \code{NULL}. if set to \code{both_significant}the
returned dataframe will only display genes that are significant at both
transcriptomics and proteomics levels. Setting the parameter to \code{either}will
return genes that are significant in at least one layer.}

\item{additional_database_gmt}{provide GMT file path for integrated pathway enrichment}
}
\value{
A list object with \code{dataframe} and \code{plot} slots
}
\description{
Compare differential expressed genes and proteins
}
\seealso{
Other Single-omic: 
\code{\link{DE_multiomics}()},
\code{\link{protein_DE_analysis}()},
\code{\link{rna_DE_analysis}()}
}
\concept{Single-omic}
