% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_omic_report.R
\name{single_omic_report}
\alias{single_omic_report}
\title{Generate a report for single omic analysis}
\usage{
single_omic_report(
  multiassay,
  report_folder_path = getwd(),
  report_file = "single_omic_report_Omix",
  slot = "ADvsControl",
  database = "Reactome_2016",
  clinical_covariates = c("PHF1", "amyloid", "PMD")
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{report_folder_path}{report_folder_path folder path to save the report.}

\item{report_file}{filename for report (without an extension).}

\item{slot}{Slot of interest}

\item{database}{Enrichment database \code{GO_Molecular_Function_2021},\code{GO_Cellular_Component_2021},
\code{GO_Biological_Process_2021}, \code{Reactome_2016} , \code{KEGG_2021_Human} , \code{MSigDB_Hallmark_2020}}

\item{clinical_covariates}{clinical covariates for table1}
}
\value{
Single omic analyses report html
}
\description{
Generate a report for single omic analysis
}
\seealso{
Other Report: 
\code{\link{multi_omic_report}()},
\code{\link{multi_omics_qc}()},
\code{\link{pathway_report}()},
\code{\link{proteomics_report}()},
\code{\link{qc_report}()}
}
\concept{Report}
