% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spls_wrapper.R
\name{spls_wrapper}
\alias{spls_wrapper}
\title{Performs sparse partial least squares regression on processed slots (rna and protein)}
\usage{
spls_wrapper(
  multiassay,
  dependent,
  ncomp = 3,
  test.keepX_rna = c(50, 60, 70, 80, 90, 100),
  test.keepX_protein = c(50, 60, 70, 80, 90, 100)
)
}
\arguments{
\item{multiassay}{Multiassay object generated by Omix}

\item{dependent}{Continuous dependent variable for sparse partial least squares}
}
\value{
spls models for processed transcriptomics and proteomics
}
\description{
Performs sparse partial least squares regression on processed slots (rna and protein)
}
\seealso{
Other Single-omic: 
\code{\link{DE_multiomics}()},
\code{\link{protein_DE_analysis}()},
\code{\link{rna_DE_analysis}()},
\code{\link{single_omic_comparisons}()},
\code{\link{wgcna_wrapper}()}
}
\concept{Single-omic}
