% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{Transcription_Factor_enrichment_supervised}
\alias{Transcription_Factor_enrichment_supervised}
\title{Transcription factor target enrichment test}
\usage{
Transcription_Factor_enrichment_supervised(
  multiassay,
  communities,
  signatures,
  TF_gmt = "Enrichr_Queries.gmt",
  direction = "negative"
)
}
\arguments{
\item{multiassay}{multiassay}

\item{communities}{communities slot generated generated by
\code{communities_network}}

\item{signatures}{signatures}

\item{TF_gmt}{GMT file with Transcription Factors and target genes}

\item{direction}{string indicating direction as "positive" or "negative"}
}
\value{
TF target enrichment in each modules
}
\description{
Transcription factor target enrichment test
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
