% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_integration.R
\name{vertical_integration}
\alias{vertical_integration}
\title{Patient specific vertical integration}
\usage{
vertical_integration(
  multiassay,
  slots = c("rna_processed", "protein_processed"),
  integration = c("MOFA", "DIABLO", "sMBPLS", "iCluster", "MEIFESTO"),
  intersect_genes = FALSE,
  ID_type = "gene_name",
  dependent = "Group",
  levels = c("Control", "Case"),
  design = c("cor", "full"),
  ncomp = 2,
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10)),
  list.keepX = list(mRNA = c(50), proteins = c(50)),
  num_factors = 10,
  time = "pseudotime",
  scale_views = TRUE,
  try.N.clust = 2:4,
  most_variable_feature = FALSE,
  ...
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{slots}{Default are processed slots}

\item{integration}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{intersect_genes}{Logical whether to keep intersecting features for
integration}

\item{design}{Design matrix (design = "full"): The strength of all
relationships between dataframes is maximised (= 1) – a “fully connected” design.
If design is set on cor, the correlation between PC1 of each dataset will be
set for the design matrix}

\item{ncomp}{Number of components in \code{DIABLO},\code{sMBPLS}}

\item{range}{List of the range of numbers of features to keep in the tuning phase.
First element must be for rna, second for proteins}

\item{try.N.clust}{number of cluster to tune in \code{iCluster}}
}
\value{
Returns an integrated object in \code{multiassay@metadata$integration} to
be used for further analysis
}
\description{
Patient specific vertical integration
}
