% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_pathway.R
\name{visualise_pathway}
\alias{visualise_pathway}
\title{visualise_pathway}
\usage{
visualise_pathway(pathways, gene_data = NULL, num_path = 10)
}
\arguments{
\item{pathways}{dataframe of enrichment results}

\item{gene_data}{LogFC}

\item{num_path}{number of pathways to plot}
}
\value{
viridis
}
\description{
visualise_pathway
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
