% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_interactive_comparison}
\alias{volcano_interactive_comparison}
\title{Builds an interactive comparison plot using plotly}
\usage{
volcano_interactive_comparison(data, log2FoldChange = 0.25)
}
\arguments{
\item{log2FoldChange}{}
}
\description{
Builds an interactive comparison plot using plotly
}
