% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_plot_limma}
\alias{volcano_plot_limma}
\title{Title}
\usage{
volcano_plot_limma(
  dt,
  log2FoldChange = 0,
  n_label = 10,
  padj = 0.05,
  ylim = c(0, 10)
)
}
\arguments{
\item{ylim}{}
}
\description{
Title
}
