% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_wrapper.R
\name{wgcna_wrapper}
\alias{wgcna_wrapper}
\title{Performs WGCNA on processed slots (rna and protein)}
\usage{
wgcna_wrapper(multiassay, dependent, covariates)
}
\arguments{
\item{multiassay}{Multiassay object generated by Omix}

\item{dependent}{Continuous dependent variable for sparse partial least squares}

\item{covariates}{clinical annotations to be correlated with modules}
}
\value{
WGCNA for processed transcriptomics and proteomics
}
\description{
Performs WGCNA on processed slots (rna and protein)
}
\seealso{
Other Single-omic: 
\code{\link{DE_multiomics}()},
\code{\link{protein_DE_analysis}()},
\code{\link{rna_DE_analysis}()},
\code{\link{single_omic_comparisons}()},
\code{\link{spls_wrapper}()}
}
\concept{Single-omic}
