% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Calls chatGPT with a specific prompt}
\usage{
ask_chatgpt(
  prompt,
  api_key = "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"
)
}
\arguments{
\item{prompt}{Character vector}

\item{api_key}{Personal API key}
}
\value{
ChatGPT answer
}
\description{
Calls chatGPT with a specific prompt
}
\seealso{
Other Helper: 
\code{\link{counts_rna}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{format_res_deseq}()},
\code{\link{format_res_limma}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
