% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction_protein.R
\name{batch_correction_protein}
\alias{batch_correction_protein}
\title{Calling different proteomics correction methods.
Internal to \code{process_protein()}}
\usage{
batch_correction_protein(matrix, batch_map, batch, correction_method)
}
\arguments{
\item{matrix}{Protein abundance matrix}

\item{batch_map}{Automatically generated in \code{process_protein()}}

\item{batch}{Batch column name}

\item{correction_method}{\code{Combat} or \code{Median_centering}}
}
\value{
Batch corrected matrix
}
\description{
Calling different proteomics correction methods.
Internal to \code{process_protein()}
}
\seealso{
Other Pre-processing: 
\code{\link{combat_correction}()},
\code{\link{filter_protein}()},
\code{\link{generate_multiassay}()},
\code{\link{median_centering_correction}()},
\code{\link{process_protein}()},
\code{\link{process_rna}()}
}
\concept{Pre-processing}
