% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{cell_type_enrichment}
\alias{cell_type_enrichment}
\title{Cell type enrichment of communities}
\usage{
cell_type_enrichment(multiassay, communities)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{communities}{communities generated from \code{communities_network()}}
}
\value{
List of results
}
\description{
Cell type enrichment of communities
}
\seealso{
Other Enrichment analysis: 
\code{\link{GSEA}()},
\code{\link{enrichment_custom}()},
\code{\link{pathway_analysis_enrichr}()}
}
\concept{Enrichment analysis}
