% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_corrplot.R
\name{circular_corrplot}
\alias{circular_corrplot}
\title{Circular heatmap of module-trait correlations and module-cell type enrichments}
\usage{
circular_corrplot(moduleTraitCor, moduleColors, cell_type, names = NULL)
}
\arguments{
\item{moduleTraitCor}{moduleTraitCor from \code{multiomics_modules()}}

\item{moduleColors}{moduleColors from \code{multiomics_modules()}}

\item{cell_type}{output from \code{cell_type_enrichment()}}

\item{names}{modulesLabels from \code{multiomics_modules()}}
}
\value{
Circular correlation plot
}
\description{
Circular heatmap of module-trait correlations and module-cell type enrichments
}
