% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_corrplot.R
\name{circular_corrplot}
\alias{circular_corrplot}
\title{Circular heatmap of module-trait correlations and module-cell type enrichments}
\usage{
circular_corrplot(moduleTraitCor, moduleColors, cell_type, names = NULL)
}
\arguments{
\item{moduleTraitCor}{moduleTraitCor from \code{multiomics_modules()}}

\item{moduleColors}{moduleColors from \code{multiomics_modules()}}

\item{cell_type}{cell type plots from \code{cell_type_enrichment()}}

\item{names}{modulesLabels from \code{multiomics_modules()}}
}
\value{
Circular correlation plot
}
\description{
Circular heatmap of module-trait correlations and module-cell type enrichments
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
