% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_heatmap.R
\name{circular_heatmap}
\alias{circular_heatmap}
\title{Title}
\usage{
circular_heatmap(
  network_communities,
  cell_type_communities,
  data = "bundle_WGCNA_rna/rna-networkConstruction-auto_new.RData"
)
}
\arguments{
\item{network_communities}{The name of the network communities}

\item{cell_type_communities}{Cell type enrichment of the community}
}
\description{
Title
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{getMoHeatmap}},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
