% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_DE.R
\name{clustering_DE_analysis}
\alias{clustering_DE_analysis}
\title{Performs differential expression between clusters}
\usage{
clustering_DE_analysis(
  normalized_data = multimodal$rna_processed,
  colData,
  dependent = "Cluster",
  levels = c("CS2", "CS1"),
  log2FoldChange = 0.5
)
}
\arguments{
\item{log2FoldChange}{}
}
\description{
Performs differential expression between clusters
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()}
}
\concept{Multi-omic integration downstream analysis}
