% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_correction.R
\name{combat_correction}
\alias{combat_correction}
\title{Proteomics normalisation batch correction using Combat}
\usage{
combat_correction(matrix, batch_map, batch)
}
\arguments{
\item{matrix}{protein abundance matrix}

\item{batch_map}{generated internally to \code{batch_correction_protein()}}
}
\value{
Combat corrected data frame
}
\description{
Proteomics normalisation batch correction using Combat
}
