% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_correction.R
\name{combat_correction}
\alias{combat_correction}
\title{Proteomics normalisation batch correction using Combat}
\usage{
combat_correction(matrix, batch_map, batch)
}
\arguments{
\item{matrix}{protein abundance matrix}

\item{batch_map}{generated internally to \code{batch_correction_protein()}}

\item{batch}{Batch levels}
}
\value{
Combat corrected data frame
}
\description{
Proteomics normalisation batch correction using Combat
}
\seealso{
Other Pre-processing: 
\code{\link{batch_correction_protein}()},
\code{\link{filter_protein}()},
\code{\link{generate_multiassay}()},
\code{\link{process_protein}()},
\code{\link{process_rna}()}
}
\concept{Pre-processing}
