% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{communities_network}
\alias{communities_network}
\title{Detects communities in multi-omics network}
\usage{
communities_network(igraph, community_detection = "louvain")
}
\arguments{
\item{igraph}{Multi-omics network}

\item{community_detection}{Available methods include \code{edge_betweeness},
\code{leading_eigen},\code{walktrap},\code{fastgreedy},\code{louvain} all available from the
\code{igraph} package.}
}
\value{
List object including \code{community_object} and \code{communities}
}
\description{
Detects communities in multi-omics network
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
