% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rna.R
\name{counts_rna}
\alias{counts_rna}
\title{Adds custom processed dataset to multiassay object}
\usage{
counts_rna(multiassay, custom_processed_df)
}
\arguments{
\item{custom_processed_df}{}
}
\description{
Adds custom processed dataset to multiassay object
}
