% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_DE_multiomics.R
\name{DE_multiomics}
\alias{DE_multiomics}
\title{Performs DE on multiomics object}
\usage{
DE_multiomics(
  multiassay,
  dependent = "diagnosis",
  covariates = c("pmi", "msex", "age_death"),
  levels = c("Control", "Mid", "Late")
)
}
\arguments{
\item{multiassay}{}

\item{dependent}{}

\item{covariates}{}

\item{levels}{}
}
\description{
Performs DE on multiomics object
}
