% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_multiomics.R
\name{DE_multiomics}
\alias{DE_multiomics}
\title{Performs DE on multiomics object from the processed slots}
\usage{
DE_multiomics(
  multiassay,
  dependent = "diagnosis",
  covariates = c("pmi", "msex", "age_death"),
  levels = c("Control", "Mid", "Late"),
  log2FoldChange = 0.2
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{dependent}{Dependent variable for the differential expression analysis,
usually the disease group variable}

\item{covariates}{Technical and biological covariates. Set as NULL is data
is denoised.}

\item{levels}{Character vector with reference group as first element.}

\item{log2FoldChange}{Log2FC absolute threshold. Default to 0.2}
}
\value{
List of differential expression results
}
\description{
Performs DE on multiomics object from the processed slots
}
\seealso{
Other Single-omic: 
\code{\link{protein_DE_analysis}()},
\code{\link{rna_DE_analysis}()},
\code{\link{single_omic_comparisons}()}
}
\concept{Single-omic}
