% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.community_graph}
\alias{.community_graph}
\title{Creates a subnetwork graph based on detected communities}
\usage{
.community_graph(igraph, community_object, community = 1)
}
\arguments{
\item{igraph}{Multi-omics network}

\item{community_object}{Community object generated by  \code{.communities_network}function}

\item{community}{Community of interest}
}
\value{
A list including \code{graph} and \code{hubs}
}
\description{
Creates a subnetwork graph based on detected communities
}
