% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{.get_ID_names}
\alias{.get_ID_names}
\title{Convert nomenclature}
\usage{
.get_ID_names(id, omic = "protein", from = "uniprot_id", to = "gene_name")
}
\arguments{
\item{to}{}
}
\description{
Convert nomenclature
}
