% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.multiomics_network}
\alias{.multiomics_network}
\title{Get multiomics networks from multiassay object}
\usage{
.multiomics_network(multiassay, list, correlation_threshold = 0.5)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{list}{named list of processed omic slots and features of interest}

\item{correlation_threshold}{Absolute correlation threshold to draw edges in the network}
}
\value{
List containing \code{graph} and \code{matrix} slots
}
\description{
Get multiomics networks from multiassay object
}
