% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{.multiomics_network_matrix}
\alias{.multiomics_network_matrix}
\title{Get multiomics networks from multimodal object}
\usage{
.multiomics_network_matrix(
  multimodal_object,
  list,
  correlation_threshold = 0.5
)
}
\arguments{
\item{correlation_threshold}{}
}
\description{
Get multiomics networks from multimodal object
}
