% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multiomic_signature.R
\name{extract_multiomic_signature}
\alias{extract_multiomic_signature}
\title{Extracts a multi-omics signature from loadings}
\usage{
extract_multiomic_signature(multiassay, integration = "DIABLO", component = 1)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{component}{Chosen component for downstream analysis. Default to 1}
}
\value{
List object
}
\description{
Extracts a multi-omics signature from loadings
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
