% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multiomic_signature.R
\name{extract_multiomic_signature}
\alias{extract_multiomic_signature}
\title{Extracts a multi-omics signature from loadings}
\usage{
extract_multiomic_signature(multiassay, integration = "DIABLO", component = 1)
}
\arguments{
\item{component}{}
}
\description{
Extracts a multi-omics signature from loadings
}
