% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multiomic_signature.R
\name{extract_multiomic_signature}
\alias{extract_multiomic_signature}
\title{Extracts a multi-omics signature from loadings}
\usage{
extract_multiomic_signature(multiassay, integration = "DIABLO", component = 1)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{component}{Chosen component for downstream analysis. Default to 1}
}
\value{
List object
}
\description{
Extracts a multi-omics signature from loadings
}
