% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{format_res_deseq}
\alias{format_res_deseq}
\title{Formatting differential expression results from DESEQ2}
\usage{
format_res_deseq(
  dt,
  log2FoldChange = 0,
  padj = 0.05,
  n_label = 10,
  gene_id_conversion = NULL,
  filter_protein_coding = TRUE
)
}
\arguments{
\item{dt}{Differential expression results dataframe from DESEQ2}

\item{log2FoldChange}{Log2FC absolute threshold}

\item{padj}{adjusted p value threshold. Default to 0.05}

\item{n_label}{maximum number of label on volcano plot}

\item{gene_id_conversion}{Default to NULL. Can provide a dataframe
with \code{gene_name} and \code{uniprot_id} columns to easily convert nomenclature.}

\item{filter_protein_coding}{Logical as whether to filter protein coding gene.
Default to TRUE.}
}
\value{
Formatted data frame
}
\description{
Formatting differential expression results from DESEQ2
}
\seealso{
Other Helper: 
\code{\link{counts_rna}()},
\code{\link{format_res_limma}()},
\code{\link{get_background}()},
\code{\link{get_metadata}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
