% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{format_res_limma}
\alias{format_res_limma}
\title{Formatting differential expression results from limma}
\usage{
format_res_limma(
  dt,
  gene_id_conversion = NULL,
  log2FoldChange = 0,
  n_label = 20,
  padj = 0.05
)
}
\arguments{
\item{dt}{Differential expression results dataframe from limma}

\item{gene_id_conversion}{Default to NULL.
Can provide a dataframe with \code{gene_name}
and \code{uniprot_id} columns to easily convert nomenclature.}

\item{log2FoldChange}{Log2FC absolute threshold}

\item{n_label}{maximum number of label on volcano plot}

\item{padj}{adjusted p value threshold. Default to 0.05}
}
\value{
Formatted data frame
}
\description{
Formatting differential expression results from limma
}
\seealso{
Other Helper: 
\code{\link{counts_rna}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{format_res_deseq}()},
\code{\link{processed_proteomics}()},
\code{\link{processed_rna}()}
}
\concept{Helper}
