% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata}
\usage{
get_metadata(multiassay, omic, slot = c("rna_processed", "protein_processed"))
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{omic}{Supported omic layers include \code{protein} or \code{rna}}

\item{slot}{Defaults to \code{c("rna_processed", "protein_processed")}}
}
\value{
colData
}
\description{
Get metadata
}
\keyword{internal}
