% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{GSEA}
\alias{GSEA}
\title{Performs functional gene set enrichment analysis}
\usage{
GSEA(gene_list, pval)
}
\arguments{
\item{gene_list}{vector of genes}

\item{pval}{threshold for adjusted pvalue}
}
\value{
List of results
}
\description{
Performs functional gene set enrichment analysis
}
\seealso{
Other Enrichment analysis: 
\code{\link{cell_type_enrichment}()},
\code{\link{enrichment_custom}()},
\code{\link{pathway_analysis_enrichr}()}
}
\concept{Enrichment analysis}
