% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_DIABLO}
\alias{integrate_with_DIABLO}
\title{Vertical integration with DIABLO including model tuning}
\usage{
integrate_with_DIABLO(
  multimodal_omics,
  Y,
  ncomp,
  design = c("cor", "full"),
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10))
)
}
\arguments{
\item{design}{Design matrix (design = "full"): The strength of all
relationships between dataframes is maximised (= 1) – a “fully connected” design.
If design is set on cor, the correlation between PC1 of each dataset will be
set for the design matrix}

\item{range}{List of the range of numbers of features to keep in the tuning phase.
First element must be for rna, second for proteins}
}
\description{
Vertical integration with DIABLO including model tuning
}
