% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_DIABLO}
\alias{integrate_with_DIABLO}
\title{Vertical integration with DIABLO including model tuning}
\usage{
integrate_with_DIABLO(
  multimodal_omics,
  Y,
  ncomp,
  design = c("cor", "full"),
  range = list(mRNA = seq(5, 10, by = 10), proteins = seq(5, 10, by = 10))
)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{.get_multimodal_object}}

\item{Y}{Outcome variable}

\item{ncomp}{Number of components}

\item{design}{Possible designs for integration are \code{cor}(strength of correlation between PC1 of omics layers)
or \code{full} (The strength of all relationships between dataframes is maximised (= 1) – a “fully connected” design)}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with DIABLO including model tuning
}
