% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_iCluster}
\alias{integrate_with_iCluster}
\title{Vertical integration with iCluster}
\usage{
integrate_with_iCluster(multimodal_omics, try.N.clust = 2:6)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{.get_multimodal_object}}

\item{try.N.clust}{Range of number of potential clusters in the optimisation step}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with iCluster
}
\seealso{
Other Multi-omic integration: 
\code{\link{getClustNum}()},
\code{\link{integrate_with_DIABLO}()},
\code{\link{integrate_with_MBPLS}()},
\code{\link{integrate_with_MEFISTO}()},
\code{\link{integrate_with_MOFA}()},
\code{\link{integrate_with_sMBPLS}()},
\code{\link{vertical_integration}()}
}
\concept{Multi-omic integration}
