% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_MEIFESTO}
\alias{integrate_with_MEIFESTO}
\title{Vertical integration with MEIFESTO}
\usage{
integrate_with_MEIFESTO(
  multimodal_omics,
  num_factors = 5,
  scale_views = TRUE,
  metadata,
  time = "pseudotime",
  ...
)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{get_multimodal_object}}

\item{num_factors}{Number of factors}

\item{scale_views}{Logical whether to scale omics layers}

\item{metadata}{metadata object generated by \code{get_multimodal_object}}

\item{time}{Time covariates}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with MEIFESTO
}
\seealso{
Other Multi-omic integration: 
\code{\link{integrate_with_DIABLO}()},
\code{\link{integrate_with_MBPLS}()},
\code{\link{integrate_with_MOFA}()},
\code{\link{integrate_with_iCluster}()},
\code{\link{integrate_with_sMBPLS}()},
\code{\link{vertical_integration}()}
}
\concept{Multi-omic integration}
