% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_MOFA}
\alias{integrate_with_MOFA}
\title{Vertical integration with MOFA}
\usage{
integrate_with_MOFA(
  multimodal_omics,
  num_factors = 5,
  scale_views = T,
  metadata,
  ...
)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{.get_multimodal_object}}

\item{num_factors}{Number of factors}

\item{scale_views}{Logical whether to scale omics layers}

\item{metadata}{}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with MOFA
}
