% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_models.R
\name{integrate_with_sMBPLS}
\alias{integrate_with_sMBPLS}
\title{Vertical integration with SMBPLS}
\usage{
integrate_with_sMBPLS(
  multimodal_omics,
  Y,
  design,
  ncomp,
  list.keepX = list(mRNA = c(50), proteins = c(50)),
  ...
)
}
\arguments{
\item{multimodal_omics}{multimodal_object generated by \code{get_multimodal_object}}

\item{Y}{Outcome variable}

\item{design}{Available values are cor, full, avg. Possible designs for
integration are \code{cor}(strength of correlation between PC1 of omics layers)
or \code{full} (The strength of all relationships between dataframes is maximised
(= 1) – a “fully connected” design)}

\item{ncomp}{Number of components}

\item{list.keepX}{grid of keepX values to keep in the tuning phase}
}
\value{
a list object with \code{multimodal_object} and \code{model} slots
}
\description{
Vertical integration with SMBPLS
}
\seealso{
Other Multi-omic integration: 
\code{\link{integrate_with_DIABLO}()},
\code{\link{integrate_with_MBPLS}()},
\code{\link{integrate_with_MEIFESTO}()},
\code{\link{integrate_with_MOFA}()},
\code{\link{integrate_with_iCluster}()},
\code{\link{vertical_integration}()}
}
\concept{Multi-omic integration}
