% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_clustering.R
\name{integrative_results_clustering}
\alias{integrative_results_clustering}
\title{Title}
\usage{
integrative_results_clustering(
  multiassay,
  intersect_genes = FALSE,
  slots,
  ID_type,
  integration,
  enrichment_method,
  annotations_continuous,
  annotations_cat,
  res.path,
  correlation_threshold = 0.5,
  geneset = NULL
)
}
\arguments{
\item{correlation_threshold}{}
}
\description{
Title
}
