% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_clustering.R
\name{integrative_results_clustering}
\alias{integrative_results_clustering}
\title{Downstream analysis for integrated multi-omics clustering}
\usage{
integrative_results_clustering(
  multiassay,
  intersect_genes = FALSE,
  slots,
  ID_type,
  integration,
  enrichment_method,
  annotations_continuous,
  annotations_cat,
  res.path,
  correlation_threshold = 0.5,
  geneset = NULL
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{correlation_threshold}{}
}
\description{
Downstream analysis for integrated multi-omics clustering
}
