% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_clustering.R
\name{integrative_results_clustering}
\alias{integrative_results_clustering}
\title{Downstream analysis for integrated multi-omics data - Clustering models}
\usage{
integrative_results_clustering(
  multiassay,
  integration_model = "iCluster",
  dependent = "diagnosis",
  sense_check_variable = "PHF1",
  covariates = c("PHF1", "amyloid", "PHF1"),
  probability_threshold = 0.8,
  correlation_threshold = 0.4,
  log2FoldChange = 1,
  community_detection = "leading_eigen",
  disease_id = "MONDO_0004975",
  TF_fp
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration_model}{Possible clustering integration methods are \code{iCluster}}

\item{dependent}{Dependent variable for groups.}

\item{sense_check_variable}{sense check}

\item{covariates}{clinical covariates}

\item{probability_threshold}{Posterior probability minimum threshold}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}

\item{log2FoldChange}{logFC cutoff}

\item{community_detection}{community detection method}

\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease) Different disease ontologies can be found on
https://www.ebi.ac.uk/ols/ontologies}

\item{TF_fp}{file path GMT file with Transcription Factors and target genes. Check \verb{https://maayanlab.cloud/chea3/}}
}
\value{
List object of integrated results
}
\description{
Downstream analysis for integrated multi-omics data - Clustering models
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{circos_TF}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
