% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{integrative_results_supervised}
\alias{integrative_results_supervised}
\title{Downstream analysis for integrated multi-omics data}
\usage{
integrative_results_supervised(
  multiassay,
  integration = "DIABLO",
  component = 1,
  correlation_threshold = 0.2,
  disease_id = "MONDO_0004975",
  enrichment_method = "enrichr",
  simplify = FALSE,
  compare_communities = FALSE
)
}
\arguments{
\item{compare_communities}{}
}
\description{
Downstream analysis for integrated multi-omics data
}
