% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_supervised.R
\name{integrative_results_supervised}
\alias{integrative_results_supervised}
\title{Downstream analysis for integrated multi-omics data}
\usage{
integrative_results_supervised(
  multiassay,
  integration = "DIABLO",
  component = 1,
  correlation_threshold = 0.2,
  disease_id = "MONDO_0004975",
  enrichment_method = "enrichr",
  simplify = FALSE,
  compare_communities = FALSE
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{integration}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{component}{Chosen component for downstream analysis. Default to 1}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}

\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease)}

\item{enrichment_method}{Available methods include \code{enrichGO} and \code{enrichr}.
Default to \code{enrichr}.}

\item{simplify}{Logical whether to simplify pathways ontologies using
\code{clusterProfiler}}

\item{compare_communities}{Logical whether to compare multi-omics network
communities using \code{compareCluster}}
}
\value{
List object of integrated results
}
\description{
Downstream analysis for integrated multi-omics data
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()}
}
\concept{Multi-omic integration downstream analysis}
