% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_unsupervised.R
\name{integrative_results_unsupervised}
\alias{integrative_results_unsupervised}
\title{Downstream analysis for integrated multi-omics data}
\usage{
integrative_results_unsupervised(
  multiassay,
  integration_model = "MOFA",
  dependent = "diagnosis",
  sense_check_variable = "PHF1",
  covariates = c("PHF1", "amyloid", "AT8"),
  weights_threshold = 1.5,
  correlation_threshold = 0.4,
  community_detection = "leading_eigen",
  pseudotime_analysis = TRUE,
  time = NA,
  disease_id = "MONDO_0004975",
  TF_fp,
  ctd
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration_model}{Possible unsupervised integration methods are \code{MOFA}}

\item{dependent}{Dependent variable for groups.}

\item{sense_check_variable}{sense check}

\item{covariates}{cov}

\item{weights_threshold}{weights threshold (+/- SD from the mean)}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}

\item{community_detection}{community detection method}

\item{pseudotime_analysis}{TRUE/ FALSE}

\item{time}{tim}

\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease) Different disease ontologies can be found on
https://www.ebi.ac.uk/ols/ontologies}

\item{TF_fp}{file path GMT file with Transcription Factors and target genes. Check \verb{https://maayanlab.cloud/chea3/}}

\item{ctd}{CellTypeData file. Generated by \code{EWCE} package}
}
\value{
List object of integrated results
}
\description{
Downstream analysis for integrated multi-omics data
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{circos_TF}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
