% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_unsupervised.R
\name{integrative_results_unsupervised}
\alias{integrative_results_unsupervised}
\title{Downstream analysis for integrated multi-omics data}
\usage{
integrative_results_unsupervised(
  multiassay,
  integration = "MOFA",
  dependent = "diagnosis",
  correlation_threshold = 0.5,
  disease_id = "MONDO_0004975"
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{integration}{Possible unsupervised integration methods are \code{MOFA},}

\item{dependent}{Dependent variable for groups.}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}

\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease) Different disease ontologies can be found on
https://www.ebi.ac.uk/ols/ontologies}
}
\value{
List object of integrated results
}
\description{
Downstream analysis for integrated multi-omics data
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{clustering_DE_analysis}()},
\code{\link{integrative_results_supervised}()},
\code{\link{multiomics_modules}()}
}
\concept{Multi-omic integration downstream analysis}
