% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{interactive_network}
\alias{interactive_network}
\title{Get interactive network from igraph object}
\usage{
interactive_network(igraph, communities = FALSE, cluster)
}
\arguments{
\item{igraph}{Multi-omics network}

\item{communities}{Logical as whether to show communities on graph}

\item{cluster}{If commmunities = TRUE, provide a community object generated
by  \code{communities_network}function}
}
\value{
plots \code{visNetwork} graph
}
\description{
Get interactive network from igraph object
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
