% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_centering_correction.R
\name{median_centering_correction}
\alias{median_centering_correction}
\title{Proteomics normalisation batch correction using median centering method}
\usage{
median_centering_correction(matrix, batch_map, batch)
}
\arguments{
\item{matrix}{protein abundance matrix}

\item{batch_map}{generated internally to \code{batch_correction_protein()}}

\item{batch}{batch covariate}
}
\value{
Median normalised data frame
}
\description{
Proteomics normalisation batch correction using median centering method
}
