% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrative_results_unsupervised.R
\name{MOFA_get_relevant_factor}
\alias{MOFA_get_relevant_factor}
\title{Selects the top factor that is most strongly associated with outcomes of interest and that has the most balanced explained variation across omic layers}
\usage{
MOFA_get_relevant_factor(MOFAobject, covariate)
}
\arguments{
\item{MOFAobject}{MOFA model}

\item{covariate}{Covariate of interest}
}
\value{
String of relevant factor
}
\description{
Selects the top factor that is most strongly associated with outcomes of interest and that has the most balanced explained variation across omic layers
}
\keyword{internal}
