% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_modules.R
\name{multiomics_modules}
\alias{multiomics_modules}
\title{Compute the module eigen value (PC1 of scaled transcriptomics/proteomics
expression) and correlates it to chosen covariates}
\usage{
multiomics_modules(
  multiassay,
  metadata,
  covariates,
  communities,
  filter_string_50 = FALSE
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{metadata}{metadata dataframe including chosen \code{covariates}}

\item{covariates}{clinical covariates}

\item{communities}{communities slot generated generated by
\code{communities_network}}

\item{filter_string_50}{should be set to TRUE if \code{MOFA} integration was used}
}
\value{
a list object with \code{metadata_modules}, \code{corrplot},\code{moduleTraitCor},
\code{moduleTraitPvalue},\code{moduleColors},\code{moduleLabels},\code{modules_eigen_value}
}
\description{
Compute the module eigen value (PC1 of scaled transcriptomics/proteomics
expression) and correlates it to chosen covariates
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{circos_TF}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{pseudotime_inference}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
