% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_modules.R
\name{multiomics_modules}
\alias{multiomics_modules}
\title{Compute the module eigen value (PC1 of scaled transcriptomics/proteomics
expression) and correlates it to chosen covariates}
\usage{
multiomics_modules(
  multiassay,
  metadata,
  covariates = c("pseudotime", "PHF1", "amyloid", "pTau"),
  communities
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{metadata}{metadata dataframe including chosen \code{covariates}}

\item{covariates}{clinical covariates}

\item{communities}{communities slot generated generated by
\code{communities_network}}
}
\value{
a list object with \code{metadata_modules}, \code{corrplot},\code{moduleTraitCor},
\code{moduleTraitPvalue},\code{moduleColors},\code{moduleLabels},\code{modules_eigen_value}
}
\description{
Compute the module eigen value (PC1 of scaled transcriptomics/proteomics
expression) and correlates it to chosen covariates
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()}
}
\concept{Multi-omic integration downstream analysis}
