% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{multiomics_network}
\alias{multiomics_network}
\title{Get multiomics networks from multiassay object}
\usage{
multiomics_network(multiassay, list, correlation_threshold = 0.5)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix
\code{generate_multiassay}function}

\item{list}{named list of processed omic slots and features of interest}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}
}
\value{
List containing \code{graph} and \code{matrix} slots
}
\description{
Get multiomics networks from multiassay object
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
