% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{multiomics_network_matrix}
\alias{multiomics_network_matrix}
\title{Get multiomics networks from multimodal object}
\usage{
multiomics_network_matrix(multimodal_object, list, correlation_threshold = 0.5)
}
\arguments{
\item{multimodal_object}{multimodal_object generated by \code{get_multimodal_object}}

\item{list}{Named list of processed omic slots and features of interest}

\item{correlation_threshold}{Absolute correlation threshold to draw edges
in the network}
}
\value{
List containing \code{graph} and \code{matrix} slots
}
\description{
Get multiomics networks from multimodal object
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
