% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{OpenTargets}
\alias{OpenTargets}
\title{Interface to OpenTargets database}
\usage{
OpenTargets(disease_id = "MONDO_0004975", size = 3000)
}
\arguments{
\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975" (Alzheimer's Disease)
Different disease ontologies can be found on https://www.ebi.ac.uk/ols/ontologies}

\item{size}{Default to 3000}
}
\value{
ggplot showing association scores from OpenTargets
}
\description{
Interface to OpenTargets database
}
