% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{plot_OpenTarget}
\alias{plot_OpenTarget}
\title{Plot Interface to OpenTargets database}
\usage{
plot_OpenTarget(disease_id = "MONDO_0004975", size = 1000, genes)
}
\arguments{
\item{disease_id}{OpenTargets disease id. Default to "MONDO_0004975"
(Alzheimer's Disease) Different disease ontologies can be found on
https://www.ebi.ac.uk/ols/ontologies}

\item{size}{Default to 3000}

\item{genes}{input gene vector}
}
\value{
ggplot showing association scores from OpenTargets
}
\description{
Plot Interface to OpenTargets database
}
\seealso{
Other Plotting: 
\code{\link{circular_corrplot}()},
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{correlation_heatmap_clusters}()},
\code{\link{correlation_heatmap_supervised}()},
\code{\link{correlation_heatmap}()},
\code{\link{getStdiz}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_heatmap}()},
\code{\link{plot_communities}()},
\code{\link{plot_components}()},
\code{\link{plot_loadings}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
