% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_visualisations.R
\name{plot_optimal_cluster}
\alias{plot_optimal_cluster}
\title{Plot optimal number of clusters from \code{getClustNum()}}
\usage{
plot_optimal_cluster(optk1, optk2, try.N.clust)
}
\arguments{
\item{optk1}{Cluster Prediction Index}

\item{optk2}{Gap statistics}
}
\value{
plot
}
\description{
Plot optimal number of clusters from \code{getClustNum()}
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
