% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_protein.R
\name{processed_proteomics}
\alias{processed_proteomics}
\title{Add custom processed proteomics dataset to multiassay}
\usage{
processed_proteomics(multiassay, custom_processed_df)
}
\arguments{
\item{custom_processed_df}{}
}
\description{
Add custom processed proteomics dataset to multiassay
}
