% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_DE_analysis.R
\name{protein_DE_analysis}
\alias{protein_DE_analysis}
\title{Performs single omic analysis of proteomic layer.}
\usage{
protein_DE_analysis(
  multiassay,
  slot = "protein_processed",
  dependent = "diagnosis",
  covariates = NULL,
  levels = c("Control", "AD"),
  log2FoldChange = 0.5,
  padj = 0.05
)
}
\arguments{
\item{multiassay}{Multiassay object generated by Omix}

\item{slot}{Default to "protein_processed"}

\item{dependent}{Dependent variable for the DeSEQ2 analysis, usually the
disease group variable}

\item{covariates}{Technical and biological covariates}

\item{levels}{Character vector with reference group as first element.}

\item{log2FoldChange}{Log2FC absolute threshold}

\item{padj}{adjusted p value threshold. Default to 0.05}
}
\value{
Adds differential expression results to Multiassay object generated
by Omix
}
\description{
Performs single omic analysis of proteomic layer.
}
\seealso{
Other Single-omic: 
\code{\link{DE_multiomics}()},
\code{\link{rna_DE_analysis}()},
\code{\link{single_omic_comparisons}()}
}
\concept{Single-omic}
