% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_inference.R
\name{pseudotime_inference}
\alias{pseudotime_inference}
\title{pseudotime_inference}
\usage{
pseudotime_inference(
  model,
  clusters,
  time_factor = 8,
  second_factor = 1,
  pseudotime_var = "pseudotime",
  start.clus = 1,
  end.clus = 3
)
}
\arguments{
\item{second_factor}{}
}
\description{
pseudotime_inference
}
