% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_inference.R
\name{pseudotime_inference}
\alias{pseudotime_inference}
\title{Pseudotime_inference based on integrated embeddings and clusters}
\usage{
pseudotime_inference(
  model,
  clusters,
  time_factor = 8,
  second_factor = 1,
  lineage = "Lineage1",
  start.clus = 1,
  end.clus = 2
)
}
\arguments{
\item{model}{MOFA model from integration in \code{multiassay@metadata$integration$MOFA}}

\item{clusters}{cluster generated from MOFA2::cluster_samples or other
clustering method}

\item{time_factor}{MOFA factor corresponding to time}

\item{second_factor}{MOFA factor used as other axis of variation in
the embedding}

\item{lineage}{lineage}

\item{start.clus}{Starting cluser}

\item{end.clus}{Ending cluser}
}
\value{
A list object with \code{model} and \code{plot} slots
}
\description{
Pseudotime_inference based on integrated embeddings and clusters
}
\seealso{
Other Multi-omic integration downstream analysis: 
\code{\link{OpenTarget_dataframe}()},
\code{\link{Transcription_Factor_enrichment_supervised}()},
\code{\link{Transcription_Factor_enrichment}()},
\code{\link{clustering_DE_analysis}()},
\code{\link{extract_loadings}()},
\code{\link{extract_multiomic_signature}()},
\code{\link{extract_weigths}()},
\code{\link{integrative_results_clustering}()},
\code{\link{integrative_results_sparse_supervised}()},
\code{\link{integrative_results_supervised}()},
\code{\link{integrative_results_unsupervised}()},
\code{\link{multiomics_modules}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_module_trajectory}()},
\code{\link{posterior_probability}()},
\code{\link{sign_signature}()}
}
\concept{Multi-omic integration downstream analysis}
