% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_inference.R
\name{pseudotime_inference}
\alias{pseudotime_inference}
\title{Pseudotime_inference based on integrated embeddings and clusters}
\usage{
pseudotime_inference(
  model,
  clusters,
  time_factor = 8,
  second_factor = 1,
  lineage = "Lineage1",
  ...
)
}
\arguments{
\item{model}{MOFA model from integration in \code{multiassay@metadata$integration$MOFA}}

\item{clusters}{cluster generated from MOFA2::cluster_samples or other
clustering method}

\item{time_factor}{MOFA factor corresponding to time}

\item{second_factor}{MOFA factor used as other axis of variation in
the embedding}
}
\value{
A list object with \code{model} and \code{plot} slots
}
\description{
Pseudotime_inference based on integrated embeddings and clusters
}
\seealso{
Other Pseudotime inference: 
\code{\link{pseudotime_ordering_modules}()}
}
\concept{Pseudotime inference}
