% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_ordering_modules.R
\name{pseudotime_ordering_modules}
\alias{pseudotime_ordering_modules}
\title{Generates a plot relating modules eigenvalue along pseudotime}
\usage{
pseudotime_ordering_modules(
  metadata,
  modules,
  variable,
  pseudotime_var = "pseudotime"
)
}
\arguments{
\item{metadata}{metadata slot generated by Omix
\code{multiomics_modules}function}

\item{modules}{character vector with module names to plot}

\item{variable}{Continuous clinical covariate}

\item{pseudotime_var}{Pseudotime covariate}
}
\value{
list of of plots - one for each module provided
}
\description{
Generates a plot relating modules eigenvalue along pseudotime
}
\seealso{
Other Pseudotime inference: 
\code{\link{pseudotime_inference}()}
}
\concept{Pseudotime inference}
