% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_omic_comparison.R
\name{single_omic_comparisons}
\alias{single_omic_comparisons}
\title{Compare differential expressed genes and proteins}
\usage{
single_omic_comparisons(
  multiassay,
  slot = "ADvsControl",
  threshold = 0.05,
  pvalue = c("adj", "pval", "all"),
  filtering_options = "both_significant"
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{slot}{If more than two comparison groups designates the slot name of
interest in \code{names(multiassay@metadata$DEG)} and
\code{names(multiassay@metadata$DEP)}}

\item{threshold}{is the pvalue threshold. Default to 0.05}

\item{pvalue}{Comparison reference.can be set to \code{pval} or \code{adj}, to consider
the normal p values or the ones after multiple testing correction,
respectively. With pvalue == \code{all}, no p value threshold is applied and all genes will be
compared. all genes are kept and only the direction is compared.}

\item{filtering_options}{default to \code{NULL}. if set to \code{both_significant}the
returned dataframe will only display genes that are significant at both
transcriptomics and proteomics levels. Setting the parameter to \code{either}will
return genes that are significant in at least one layer.}
}
\value{
A list object with \code{dataframe} and \code{plot} slots
}
\description{
Compare differential expressed genes and proteins
}
\seealso{
Other Single-omic: 
\code{\link{protein_DE_analysis}()},
\code{\link{rna_DE_analysis}()}
}
\concept{Single-omic}
