% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_integration.R
\name{vertical_integration}
\alias{vertical_integration}
\title{Patient specific vertical integration}
\usage{
vertical_integration(
  multiassay,
  slots = c("rna_processed", "protein_processed"),
  integration,
  intersect_genes = FALSE,
  ID_type = "gene_name",
  dependent = "Group",
  levels = c("Control", "Case"),
  design = "cor",
  ncomp = 2,
  range = list(mRNA = seq(5, 100, by = 10), proteins = seq(5, 100, by = 10)),
  list.keepX = list(mRNA = c(50), proteins = c(50)),
  num_factors = 10,
  time = "pseudotime",
  scale_views = TRUE,
  try.N.clust = 2:4,
  most_variable_feature = FALSE
)
}
\arguments{
\item{multiassay}{Multiassay experiment object generated by Omix}

\item{slots}{Default are processed slots}

\item{integration}{Possible integration methods are \code{MOFA},
\code{DIABLO},\code{sMBPLS},\code{iCluster},\code{MEIFESTO}}

\item{intersect_genes}{Logical whether to keep intersecting features for
integration}

\item{ID_type}{Default to gene_name}

\item{dependent}{Dependent variable for the DeSEQ2 analysis, usually the
disease group variable}

\item{levels}{Character vector with reference group as first element.
Set parameter as NULL if dependent is NULL.}

\item{design}{Available options "cor" and "full". Default is "cor".
Design matrix (design = "full"): The strength of all
relationships between dataframes is maximised (= 1) – a “fully connected” design.
If design is set on cor, the correlation between PC1 of each dataset will be
set for the design matrix}

\item{ncomp}{Number of components in \code{DIABLO},\code{sMBPLS}}

\item{range}{List of the range of numbers of features to keep in the tuning
phase. First element must be for rna, second for proteins}

\item{list.keepX}{if integration == \code{sMBPLS}, number of features to keep per view.}

\item{num_factors}{if integration == \code{MOFA}, number of factors. Default to 10.}

\item{time}{Pseudotime covariate}

\item{scale_views}{Logical whether to scale omic views}

\item{try.N.clust}{number of cluster to tune in \code{iCluster}}

\item{most_variable_feature}{Logical whether to keep most variable features}
}
\value{
Returns an integrated object in \code{multiassay@metadata$integration} to
be used for further analysis
}
\description{
Patient specific vertical integration
}
\seealso{
Other Multi-omic integration: 
\code{\link{getClustNum}()},
\code{\link{integrate_with_DIABLO}()},
\code{\link{integrate_with_MBPLS}()},
\code{\link{integrate_with_MEIFESTO}()},
\code{\link{integrate_with_MOFA}()},
\code{\link{integrate_with_iCluster}()},
\code{\link{integrate_with_sMBPLS}()}
}
\concept{Multi-omic integration}
