% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_interactive}
\alias{volcano_interactive}
\title{Builds an interactive volcano plot using plotly}
\usage{
volcano_interactive(data, log2FoldChange = 0.25)
}
\arguments{
\item{log2FoldChange}{}
}
\description{
Builds an interactive volcano plot using plotly
}
