% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_interactive}
\alias{volcano_interactive}
\title{Builds an interactive volcano plot using plotly}
\usage{
volcano_interactive(data, log2FoldChange = 0.25, padj = 0.05)
}
\arguments{
\item{data}{Formatted Differential expression results from
\code{format_res_deseq()} or \code{format_res_limma()}}

\item{log2FoldChange}{Log2FC absolute threshold. Default to 0.25}

\item{padj}{adjusted p value threshold. Default to 0.05}
}
\value{
Interactive volcano plot
}
\description{
Builds an interactive volcano plot using plotly
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
