% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_interactive_comparison}
\alias{volcano_interactive_comparison}
\title{Builds an interactive comparison plot using plotly}
\usage{
volcano_interactive_comparison(data)
}
\arguments{
\item{data}{Formatted Comparative Differential expression results from \code{single_omic_comparisons()}}
}
\value{
Interactive plot
}
\description{
Builds an interactive comparison plot using plotly
}
