% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_interactive_comparison}
\alias{volcano_interactive_comparison}
\title{Builds an interactive comparison plot using plotly}
\usage{
volcano_interactive_comparison(data)
}
\arguments{
\item{data}{Formatted Comparative Differential expression results
from \code{single_omic_comparisons()}}
}
\value{
Interactive plot
}
\description{
Builds an interactive comparison plot using plotly
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()},
\code{\link{volcano_plot_limma}()}
}
\concept{Plotting}
