% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_plot_deseq}
\alias{volcano_plot_deseq}
\title{Builds a volcano plot from DESEQ2 results}
\usage{
volcano_plot_deseq(
  dt,
  log2FoldChange = 0,
  n_label = 10,
  padj = 0.05,
  ylim = c(0, 10)
)
}
\arguments{
\item{ylim}{}
}
\description{
Builds a volcano plot from DESEQ2 results
}
