% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_formating.R
\name{volcano_plot_limma}
\alias{volcano_plot_limma}
\title{Builds a volcano plot from limma results}
\usage{
volcano_plot_limma(dt, log2FoldChange = 0, padj = 0.05)
}
\arguments{
\item{dt}{Formatted Differential expression results from \code{format_res_limma()}}

\item{log2FoldChange}{Log2FC absolute threshold}

\item{padj}{adjusted p value threshold. Default to 0.05}
}
\value{
Volcano plot
}
\description{
Builds a volcano plot from limma results
}
\seealso{
Other Plotting: 
\code{\link{communities_network}()},
\code{\link{community_graph}()},
\code{\link{interactive_network}()},
\code{\link{multiomics_network_cluster}()},
\code{\link{multiomics_network_matrix}()},
\code{\link{multiomics_network}()},
\code{\link{plot_OpenTarget}()},
\code{\link{plot_optimal_cluster}()},
\code{\link{volcano_interactive_comparison}()},
\code{\link{volcano_interactive}()},
\code{\link{volcano_plot_deseq}()}
}
\concept{Plotting}
